fix dto and domain model, automapping, fix controller functions

+++ Data/Entities/Brand.cs
@@ -6,7 +6,7 @@ namespace api.fernflowers.com.Data.Entities;
 
 public class Brand
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Name { get; set; }
-    public int VaccineId { get; set; }
+    public long VaccineId { get; set; }
 }
